Renaming the S3 component as it was having issues.
